refactor: migrate file to comp
