+++ app/application/history/page.tsx
@@ -2,10 +2,10 @@ import { getTotalClaimableRewards } from "@/app/actions/history";
 import History from "@/app/components/History/History";
 import HistoryHeader from "@/app/components/HistoryHeader/HistoryHeader";
 import ProfileNavigation from "@/app/components/ProfileNavigation/ProfileNavigation";
-import { ReopenMysteryBox } from "@/app/components/ReopenMysteryBox/ReopenMysteryBox";
 import { getAllQuestionsReadyForReveal } from "@/app/queries/history";
 import { getUnopenedMysteryBox } from "@/app/queries/mysteryBox";
 import { getProfileImage } from "@/app/queries/profile";
+import { ReopenMysteryBox } from "@/components/MysteryBox/ReopenMysteryBox";
 
 export default async function Page() {
   const [revealableQuestions, totalClaimableRewards, profileImg, mysteryBox] =
