+++ components/MysteryBox/ReopenMysteryBox.tsx
